refactor(diffshub): normalize theme provider filenames
