+++ apps/diffshub/components/RootLayout.tsx
@@ -4,7 +4,7 @@ import type { ReactNode } from 'react';
 
 import { PreloadHighlighter } from '@/components/PreloadHighlighter';
 import { ScrollbarGutterVariables } from '@/components/ScrollbarGutterVariables';
-import { ThemeProvider } from '@/components/theme-provider';
+import { ThemeProvider } from '@/components/ThemeProvider';
 import { Toaster } from '@/components/Toaster';
 import { WorkerPoolContext } from '@/components/WorkerPoolContext';
 
@@ -40,7 +40,7 @@ const themeBootstrapScript = `(${String(function applyInitialTheme() {
     // Set the iOS navbar tint before first paint so it matches the resolved
     // mode immediately. The meta is created here (not authored in JSX, which
     // React 19 would hoist into a duplicate) and owned by JS thereafter.
-    // Literals mirror MODE_THEME_COLOR in theme-provider.tsx (this stringified
+    // Literals mirror SCHEME_THEME_COLOR in ThemeProvider.tsx (this stringified
     // script can't import it); keep them in sync.
     let themeColorMeta = document.querySelector('meta[name="theme-color"]');
     if (themeColorMeta == null) {
